+++ flows/project_create.py
@@ -130,24 +130,31 @@ class Flow(BaseLinearFlow):
             )
         )
 
-        # Add inherited owners (list of usernames)
-        for i_owner in self.flow_data.get('inherited_owners') or []:
+        # Add inherited owners
+        for username in set(
+            [r['username'] for r in self.flow_data.get('roles_add', [])]
+        ):
             self.add_task(
                 irods_tasks.CreateUserTask(
-                    name='Create user for inherited category owner "{}"'.format(
-                        i_owner
-                    ),
+                    name='Create user "{}" in irods'.format(username),
                     irods=self.irods,
-                    inject={'user_name': i_owner, 'user_type': 'rodsuser'},
+                    inject={'user_name': username, 'user_type': 'rodsuser'},
                 )
             )
 
+        for role_add in self.flow_data.get('roles_add', []):
+            project_group = get_project_group_name(role_add['project_uuid'])
+
             self.add_task(
                 irods_tasks.AddUserToGroupTask(
-                    name='Add inherited owner "{}" to project user '
-                    'group'.format(i_owner),
+                    name='Add user "{}" to project user group "{}"'.format(
+                        role_add['username'], project_group
+                    ),
                     irods=self.irods,
-                    inject={'group_name': project_group, 'user_name': i_owner},
+                    inject={
+                        'group_name': project_group,
+                        'user_name': role_add['username'],
+                    },
                 )
             )